Add Papas Detector service with CMS implemented
