--- Sim/PapasDetInterface/CMakeLists.txt
@@ -1,6 +1,0 @@
-################################################################################
-# Package: PapasDetInterface
-################################################################################
-gaudi_subdir(PapasDetInterface v1r0)
-
-gaudi_install_headers(PapasDetInterface)
