+++ Sim/SimPapas/src/PapasAlg.cpp
@@ -1,31 +1,30 @@
 #include "PapasAlg.h"
 
+#include "SimPapasDetector/IPapasDetSvc.h"
 #include "SimPapas/IPapasTool.h"
-#include "PapasDetInterface/IPapasDetSvc.h"
+#include "papas/detectors/Detector.h"
 #include "papas/utility/PDebug.h"
 #include "papas/utility/TRandom.h"
-#include "papas/detectors/Detector.h"
-
 
 #include <iostream>
 DECLARE_COMPONENT(PapasAlg)
 
 PapasAlg::PapasAlg(const std::string& name, ISvcLocator* svcLoc) : GaudiAlgorithm(name, svcLoc) {
   declareProperty("tools", m_toolNames);
-  
   m_eventno = 0;
 }
 
 StatusCode PapasAlg::initialize() {
-    for (auto& toolname : m_toolNames) {
+  for (auto& toolname : m_toolNames) {
     m_tools.push_back(tool<IPapasTool>(toolname));
   }
-  m_papasDetSvc = service("CMSDetSvc");
-  //, "m_cyl1" = 1.3, "m_cyl2" = 1.55
+  m_papasDetSvc = service(m_detServiceName);
   if (!m_papasDetSvc) {
-    error() << "Unable to locate Geant Simulation Service" << endmsg;
+    error() << "Unable to locate Papas Detector Service " << m_detServiceName << endmsg;
     return StatusCode::FAILURE;
   }
+  m_spDetector = m_papasDetSvc->detector();
+  
   if (m_physicsDebugFile != "") {
     debug() << "Papas: physics output " << m_physicsDebugFile << endmsg;
     papas::PDebug::File(m_physicsDebugFile);
@@ -48,10 +47,6 @@ StatusCode PapasAlg::execute() {
   pevent.setEventNo(m_eventno);
   papas::PDebug::write("Event: {}", m_eventno);
   m_eventno++;
-  
-  
-  //std::shared_ptr<papas::Detector> det;
-  auto det = m_papasDetSvc->detector();
 
   // run the tools
   try {
@@ -62,7 +57,7 @@ StatusCode PapasAlg::execute() {
     }
     for (auto tool : m_tools) {
       // run each tool
-      tool->run(pevent, det);
+      tool->run(pevent, m_spDetector);
     }
     // summary details of what the event contains
     debug() << "PAPAS event contains " << std::endl << pevent.info() << endmsg;
