+++ Sim/SimPapas/src/PapasMergeClustersTool.cpp
@@ -1,5 +1,5 @@
 #include "PapasMergeClustersTool.h"
-
+// PAPAS
 #include "papas/graphtools/EventRuler.h"
 #include "papas/reconstruction/MergeClusters.h"
