+++ Sim/SimPapas/src/PapasMergeClustersTool.h
@@ -4,7 +4,7 @@
 #include "FWCore/DataHandle.h"
 #include "GaudiAlg/GaudiTool.h"
 #include "SimPapas/IPapasTool.h"
-
+// PAPAS
 #include "papas/datatypes/DefinitionsCollections.h"
 
 #include <string>
